+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added a multi-tenant mode for spraying multiple domains
 * Added dynamic password generation in multi-tenant mode
 * Added spray backoff with lockout thresholds
+* Storing spray output in user home directory under .spray365/output/